battery-sensor/broker/app/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
paho-mqtt>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
